# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrendPilot AI API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Trend {
  id: number;
  name: string;
  category: string;
  /** Trend popularity score 0-100 */
  score: number;
  /** Percentage change (positive = rising, negative = falling) */
  velocity: number;
  /** rising | peaking | declining | emerging */
  status: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  source?: string | null;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export interface TrendInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  velocity: number;
  status: string;
  description?: string;
  source?: string;
  tags?: string[];
}

export interface TrendUpdate {
  /** @minLength 1 */
  name?: string;
  category?: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score?: number;
  velocity?: number;
  status?: string;
  description?: string;
  source?: string;
  tags?: string[];
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface TrendSummary {
  total: number;
  rising: number;
  peaking: number;
  declining: number;
  emerging: number;
  byCategory: CategoryCount[];
  avgScore: number;
}

export interface WatchlistItem {
  id: number;
  trendId: number;
  /** @nullable */
  notes?: string | null;
  alertEnabled?: boolean;
  createdAt: string;
  trend: Trend;
}

export interface WatchlistItemInput {
  trendId: number;
  notes?: string;
  alertEnabled?: boolean;
}

export type ListTrendsParams = {
category?: string;
status?: string;
search?: string;
};

export type GetTopTrendsParams = {
limit?: number;
};