# lib/api-zod/src/generated/types/trend.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrendPilot AI API
 * OpenAPI spec version: 0.1.0
 */

export interface Trend {
  id: number;
  name: string;
  category: string;
  /** Trend popularity score 0-100 */
  score: number;
  /** Percentage change (positive = rising, negative = falling) */
  velocity: number;
  /** rising | peaking | declining | emerging */
  status: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  source?: string | null;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}