# lib/api-zod/src/generated/types/trendInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrendPilot AI API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  category: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  velocity: number;
  status: string;
  description?: string;
  source?: string;
  tags?: string[];
}