# lib/api-zod/src/generated/types/trendSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrendPilot AI API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';

export interface TrendSummary {
  total: number;
  rising: number;
  peaking: number;
  declining: number;
  emerging: number;
  byCategory: CategoryCount[];
  avgScore: number;
}