# lib/api-zod/src/generated/types/watchlistItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrendPilot AI API
 * OpenAPI spec version: 0.1.0
 */
import type { Trend } from './trend';

export interface WatchlistItem {
  id: number;
  trendId: number;
  /** @nullable */
  notes?: string | null;
  alertEnabled?: boolean;
  createdAt: string;
  trend: Trend;
}