# lib/api-zod/src/generated/types/watchlistItemInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * TrendPilot AI API
 * OpenAPI spec version: 0.1.0
 */

export interface WatchlistItemInput {
  trendId: number;
  notes?: string;
  alertEnabled?: boolean;
}